Refactor rpc_server.py into focused modules
Split the 926-line addon/FreeCADMCP/rpc_server/rpc_server.py into smaller
modules, each responsible for one concern. The XML-RPC wire contract
(method names, parameter names, return-dict shapes) is preserved exactly,
so src/freecad_mcp/freecad_client.py keeps working unchanged.

New modules under addon/FreeCADMCP/rpc_server/:
- settings.py        Settings persistence (load/save freecad_mcp_settings.json)
- ip_filter.py       FilteredXMLRPCServer + validate_allowed_ips
- gui_dispatch.py    Queue plumbing + process_gui_tasks (with the resilience
                     fix: exceptions and None returns no longer kill the
                     dispatch loop or starve the response queue)
- view_manager.py    Screenshot orientation dispatch (dict map instead of
                     11-branch if/elif) + size resolution + saveImage flow
- property_mapper.py Object dataclass + set_object_property
- object_factory.py  create_object_gui split into FEM-mesh / FEM-typed /
                     generic helpers
- fem_executor.py    CalculiX-driven FEM analysis execution
- commands.py        Five Qt Command classes + checkable-menu sync timer

rpc_server.py now keeps only the FreeCADRPC class (thin forwarders),
start_rpc_server, stop_rpc_server, and re-exports of public symbols
(load_settings, validate_allowed_ips, etc.) so addon/FreeCADMCP/InitGui.py
keeps resolving rpc_server.load_settings / rpc_server.start_rpc_server.

stop_rpc_server now posts a sentinel to the GUI dispatch queue before
shutting down the XML-RPC server, so process_gui_tasks exits cleanly
without leaving an orphan QTimer chain.

No new dependencies; all 9 modules pass ast.parse.
