+++ addon/FreeCADMCP/rpc_server/property_mapper.py
@@ -1,0 +1,92 @@
+"""Property assignment from JSON-friendly dicts onto FreeCAD document objects."""
+
+from dataclasses import dataclass, field
+from typing import Any
+
+import FreeCAD
+
+
+@dataclass
+class Object:
+    name: str
+    type: str | None = None
+    analysis: str | None = None
+    properties: dict[str, Any] = field(default_factory=dict)
+
+
+def set_object_property(
+    doc: FreeCAD.Document, obj: FreeCAD.DocumentObject, properties: dict[str, Any]
+):
+    for prop, val in properties.items():
+        try:
+            if prop in obj.PropertiesList:
+                if prop == "Placement" and isinstance(val, dict):
+                    if "Base" in val:
+                        pos = val["Base"]
+                    elif "Position" in val:
+                        pos = val["Position"]
+                    else:
+                        pos = {}
+                    rot = val.get("Rotation", {})
+                    placement = FreeCAD.Placement(
+                        FreeCAD.Vector(
+                            pos.get("x", 0),
+                            pos.get("y", 0),
+                            pos.get("z", 0),
+                        ),
+                        FreeCAD.Rotation(
+                            FreeCAD.Vector(
+                                rot.get("Axis", {}).get("x", 0),
+                                rot.get("Axis", {}).get("y", 0),
+                                rot.get("Axis", {}).get("z", 1),
+                            ),
+                            rot.get("Angle", 0),
+                        ),
+                    )
+                    setattr(obj, prop, placement)
+
+                elif isinstance(getattr(obj, prop), FreeCAD.Vector) and isinstance(
+                    val, dict
+                ):
+                    vector = FreeCAD.Vector(
+                        val.get("x", 0), val.get("y", 0), val.get("z", 0)
+                    )
+                    setattr(obj, prop, vector)
+
+                elif prop in ["Base", "Tool", "Source", "Profile"] and isinstance(
+                    val, str
+                ):
+                    ref_obj = doc.getObject(val)
+                    if ref_obj:
+                        setattr(obj, prop, ref_obj)
+                    else:
+                        raise ValueError(f"Referenced object '{val}' not found.")
+
+                elif prop == "References" and isinstance(val, list):
+                    refs = []
+                    for ref_name, face in val:
+                        ref_obj = doc.getObject(ref_name)
+                        if ref_obj:
+                            refs.append((ref_obj, face))
+                        else:
+                            raise ValueError(f"Referenced object '{ref_name}' not found.")
+                    setattr(obj, prop, refs)
+
+                else:
+                    setattr(obj, prop, val)
+            # ShapeColor is a property of the ViewObject
+            elif prop == "ShapeColor" and isinstance(val, (list, tuple)):
+                setattr(obj.ViewObject, prop, (float(val[0]), float(val[1]), float(val[2]), float(val[3])))
+
+            elif prop == "ViewObject" and isinstance(val, dict):
+                for k, v in val.items():
+                    if k == "ShapeColor":
+                        setattr(obj.ViewObject, k, (float(v[0]), float(v[1]), float(v[2]), float(v[3])))
+                    else:
+                        setattr(obj.ViewObject, k, v)
+
+            else:
+                setattr(obj, prop, val)
+
+        except Exception as e:
+            FreeCAD.Console.PrintError(f"Property '{prop}' assignment error: {e}\n")
