+++ addon/FreeCADMCP/rpc_server/settings.py
@@ -1,0 +1,43 @@
+"""Persistence of MCP RPC server settings under FreeCAD's user app data dir."""
+
+import json
+import os
+
+import FreeCAD
+
+
+_SETTINGS_FILENAME = "freecad_mcp_settings.json"
+
+_DEFAULT_SETTINGS = {
+    "remote_enabled": False,
+    "allowed_ips": "127.0.0.1",
+    "auto_start_rpc": False,
+}
+
+
+def _get_settings_path():
+    return os.path.join(FreeCAD.getUserAppDataDir(), _SETTINGS_FILENAME)
+
+
+def load_settings():
+    path = _get_settings_path()
+    if os.path.exists(path):
+        try:
+            with open(path, "r") as f:
+                settings = json.load(f)
+            for key, value in _DEFAULT_SETTINGS.items():
+                if key not in settings:
+                    settings[key] = value
+            return settings
+        except Exception as e:
+            FreeCAD.Console.PrintWarning(f"Failed to load MCP settings: {e}\n")
+    return dict(_DEFAULT_SETTINGS)
+
+
+def save_settings(settings):
+    path = _get_settings_path()
+    try:
+        with open(path, "w") as f:
+            json.dump(settings, f, indent=2)
+    except Exception as e:
+        FreeCAD.Console.PrintError(f"Failed to save MCP settings: {e}\n")
