+++ addon/FreeCADMCP/rpc_server/view_manager.py
@@ -1,0 +1,97 @@
+"""Active-view orientation, sizing, and screenshot capture."""
+
+from typing import Any
+
+import FreeCAD
+import FreeCADGui
+
+from .gui_dispatch import _flush_gui_events
+
+
+_VIEW_DISPATCH = {
+    "Isometric": "viewIsometric",
+    "Front": "viewFront",
+    "Top": "viewTop",
+    "Right": "viewRight",
+    "Back": "viewBack",
+    "Left": "viewLeft",
+    "Bottom": "viewBottom",
+    "Dimetric": "viewDimetric",
+    "Trimetric": "viewTrimetric",
+}
+
+
+def _get_view_size(view: Any) -> tuple[int, int]:
+    try:
+        size = view.getSize()
+        if isinstance(size, (list, tuple)) and len(size) >= 2:
+            return max(1, int(size[0])), max(1, int(size[1]))
+        return max(1, int(size.width())), max(1, int(size.height()))
+    except Exception:
+        return 1024, 768
+
+
+def _resolve_screenshot_size(
+    view: Any,
+    width: int | None,
+    height: int | None,
+) -> tuple[int, int]:
+    view_width, view_height = _get_view_size(view)
+    resolved_width = view_width if width is None else max(1, int(width))
+    resolved_height = view_height if height is None else max(1, int(height))
+    return resolved_width, resolved_height
+
+
+def apply_view_orientation(view: Any, view_name: str) -> None:
+    method_name = _VIEW_DISPATCH.get(view_name)
+    if method_name is None:
+        raise ValueError(f"Invalid view name: {view_name}")
+    getattr(view, method_name)()
+
+
+def save_active_screenshot(
+    save_path: str,
+    view_name: str = "Isometric",
+    width: int | None = None,
+    height: int | None = None,
+    focus_object: str | None = None,
+):
+    """Save a PNG of the active view to ``save_path``.
+
+    Returns ``True`` on success, or an error string on failure (preserves the
+    legacy GUI-handler return contract).
+    """
+    try:
+        view = FreeCADGui.ActiveDocument.ActiveView
+        if not hasattr(view, "saveImage"):
+            return "Current view does not support screenshots"
+
+        apply_view_orientation(view, view_name)
+
+        focused_selection = False
+
+        if focus_object:
+            doc = FreeCAD.ActiveDocument
+            obj = doc.getObject(focus_object) if doc else None
+            if obj:
+                FreeCADGui.Selection.clearSelection()
+                FreeCADGui.Selection.addSelection(obj)
+                FreeCADGui.SendMsgToActiveView("ViewSelection")
+                focused_selection = True
+                _flush_gui_events()
+                FreeCADGui.Selection.clearSelection()
+            else:
+                view.fitAll()
+        else:
+            view.fitAll()
+
+        _flush_gui_events()
+        resolved_width, resolved_height = _resolve_screenshot_size(view, width, height)
+        view.saveImage(save_path, resolved_width, resolved_height, "Current")
+
+        if focused_selection:
+            FreeCADGui.Selection.clearSelection()
+            _flush_gui_events(delay_ms=0)
+        return True
+    except Exception as e:
+        return str(e)
